pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "redis>=5.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-redis>=0.1.0",
    "langgraph>=0.2.0",
    "pydantic-settings>=2.5.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "ollama>=0.6.1",
]
